feat: move test user utils to common.ts
